Update style.css
The hidden class is the display block. When the condition that all fields are filled in is met, a hidden class and its child appear on which it is written that the result has not been resolved, and I put the restart button in absolute and z-index 10 so that I could restart the game.